src/app/components/list-users/list-users.component.spec.ts
import { RouterTestingModule } from '@angular/router/testing';
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { ListUsersComponent } from './list-users.component';
import { FormsModule, ReactiveFormsModule } from '@angular/forms';
import { HttpClientTestingModule } from '@angular/common/http/testing';

describe('ListUsersComponent', () => {
  let component: ListUsersComponent;
  let fixture: ComponentFixture<ListUsersComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ListUsersComponent],
      imports: [FormsModule, ReactiveFormsModule, HttpClientTestingModule, RouterTestingModule]
    })
      .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(ListUsersComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    component.listUser();
    expect(component).toBeTruthy();
  });

  it('must have a text label', () => {
    const fixture = TestBed.createComponent(ListUsersComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('.col-lg-3 label').textContent).toContain('Filter by name');
  });

  it('must have a button delete', () => {
    const fixture = TestBed.createComponent(ListUsersComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('.btn-danger').textContent).toContain('Delete');
  });

  it('must have a button delete disable', () => {
    const fixture = TestBed.createComponent(ListUsersComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('.btn-danger').disabled).toBeTruthy();
  });

  it('must have a button download', () => {
    const fixture = TestBed.createComponent(ListUsersComponent);
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('.btn-success').textContent).toContain('Download');
  });
});
